Move outfit cover into separate dynamic component
