+++ src/components/Outfits/Outfits.js
@@ -3,6 +3,7 @@ import add from '../../images/add.png';
 import { Link } from 'react-router-dom';
 import { getData } from '../../apiCalls';
 import ErrorMessage from "../ErrorMessage/ErrorMessage";
+import OutfitCover from './OutfitCover/OutfitCover';
 import './Outfits.css'
 
 const Outfits = ({setAppError, closeMenu, user, appError}) => {
@@ -27,41 +28,12 @@ const Outfits = ({setAppError, closeMenu, user, appError}) => {
       return () => setAppError(null)
   },[])
 
-  const outfitPieceImgs = (outfit) => {
-    const allImages = outfit.outfitPieces.map(piece => {
-      return <img key={piece.id} className='piece-img' src={piece.image} />
-    })
-    return allImages.slice(0, 4)
-  }
-
-    const userOutfitImages = outfits?.map(outfit => {
-      if(outfit.outfit.fullOutfitImage) {
-        return (
-          <Link key={outfit.outfit.id} to={`/outfitdetails/${outfit.outfit.id}`}>
-            <div className='outfit-img-container' > 
-              <img className='full-outfit-img' src={outfit.outfit.fullOutfitImage} alt={`image of outfit ${outfit.id}`}/>
-              <div className="overlay"></div>
-              <div className='view-outfit-btn'>View my outfit</div>
-            </div>
-          </Link>
-        )
-      } else {
-        return (
-            <Link key={outfit.outfit.id} to={`/outfitdetails/${outfit.outfit.id}`}>
-            <div className='outfit-img-container'>
-              {outfitPieceImgs(outfit)}
-              <div className="overlay"></div>
-              <div className='view-outfit-btn'>View my outfit</div>
-            </div>
-          </Link>
-        )
-      }
-    })
-
+  const userOutfitImages = outfits?.map(outfit => {
+    return <OutfitCover outfit={outfit} />
+  })
 
   return (
     <>
-    {/* <div className='outfits-page'> */}
     {appError && <ErrorMessage appError={appError}/>}
     {user ?
       <div className='my-outfits-container'>
@@ -72,7 +44,6 @@ const Outfits = ({setAppError, closeMenu, user, appError}) => {
         </div>
       </div>
       : <p>Please Login to Continue</p>}
-    {/* </div> */}
     </>
   )
 }
